DRA: rename pkg/cm/dra/plugin files
